Allow `Conjugate` option of ComplexConjugate to have rules
Prior to this, `Conjugate` could only handle a list of elements which must match
exactly for them to be complex conjugated.  This now allows for custom rules to
be allowed.

For example, the following is supported

```
Conjugate -> {
  Lambda,
  y[i_, j_] :> Conjugate[y[i, j]],
  h[i_, j_] :> h[j, i],
  z -> -z
}
```

which will leave the three rules unchanged, and implicitly understand that
`Lambda -> Conjugate[Lambda]`.

Signed-off-by: JP-Ellis <josh@jpellis.me>